Rename LPPaver.Decide to LPPaver.Algorithm
+ LPPaver.Decide.Algorithm is now LPPaver.Decide.DNF
+ Other modules LPPaver.Decide.Name have become LPPaver.Algorithm.Name
